fix(pdf): restrict outbound network access during PDF rendering
